[ios] Reset user tracking, respect content inset in resetPosition

+++ platform/darwin/MGLMapCamera.mm
@@ -43,7 +43,7 @@
 {
     return [[self alloc] initWithCenterCoordinate:centerCoordinate
                                          altitude:distance
-                                            pitch:(CGFloat)pitch
+                                            pitch:pitch
                                           heading:heading];
 }
